Add Message and MessagesRetrieve models; update get_messages to use token from Authorization header
