+++ basemodels.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Optional
+from typing import List, Optional
 from pydantic import BaseModel
 
 class TokenRequest(BaseModel):
@@ -57,3 +57,28 @@ class ServerCreate(BaseModel):
     name: str
     description: str
     owner_id: int
+
+class Message(BaseModel):
+    message: str
+    user_token: str
+    room_id: int
+    is_private: bool
+    reply_to: Optional[int] = None
+    attachments: Optional[List[str]] = None  # List of file URLs or filenames
+
+    class Config:
+        from_attributes = True
+
+class MessageResponse(BaseModel):
+    message: str
+    room_id: int
+    is_private: bool
+    reply_to: Optional[int]
+    user_id: int
+    timestamp: datetime
+    attachments: Optional[List[str]] = None
+    _id: str
+
+class MessagesRetrieve(BaseModel):
+    room_id: int
+    user_token: str
